Fix goodmemref so it works again with -D+9e and non-zero mem_base.
git-svn-id: https://dosemu.svn.sourceforge.net/svnroot/dosemu/trunk@1980 7758b857-130e-0410-ab9c-d8735a528d18

+++ src/emu-i386/simx86/emu86.h
@@ -643,7 +643,7 @@ extern int eTimeCorrect;
 //
 extern unsigned long eTSSMASK;
 extern int Running;		/* into interpreter loop */
-extern long mMaxMem;
+extern unsigned int mMaxMem;
 extern int TryMemRef;
 extern int UseLinker;
 //
